@@ -1,6 +1,6 @@
 module github.com/deloz/sid
 
-go 1.22.4
+go 1.24.4
 
 require github.com/stretchr/testify v1.10.0
 
